Refactor HAL liftover script into Python lib
