+++ pyproject.toml
@@ -7,7 +7,7 @@ build-backend = "setuptools.build_meta"
 
 [tool.pytest.ini_options]
 # CITest project is on hold and it needs to be updated before resuming its unit tests
-addopts = "-v --tb=native --server=mysql://ensadmin:$ENSADMIN_PSW@mysql-ens-compara-prod-1:4485/ --ignore=src/python/tests/test_citest.py"
+addopts = "-v --tb=native --server=mysql://ensadmin:$ENSADMIN_PSW@mysql-ens-compara-prod-1:4485/ --ignore=src/python/tests/test_citest.py --ignore=src/python/tests/helpers.py"
 testpaths = ["src/python/tests"]
 
 [tool.coverage.run]
@@ -29,4 +29,10 @@ disable = [
     "too-many-statements",
     "unspecified-encoding",
     "wildcard-import"
+]
+
+[tool.pylint.typecheck]
+signature-mutators = [
+    "click.decorators.argument",
+    "click.decorators.option",
 ]
