+++ requirements.txt
@@ -1,5 +1,6 @@
 ensembl-py@git+https://github.com/Ensembl/ensembl-py.git@1.0.0#egg=ensembl-py
 biopython>=1.76
+click>=8.1.4
 CrossMap>=0.6.5
 ete3>=3.1.1
 lxml>=4.9.2
