+++ src/python/lib/ensembl/compara/utils/tools.py
@@ -22,9 +22,49 @@ Typical usage examples::
 
 """
 
-__all__ = ['to_list']
+__all__ = ["import_module_from_file", "to_list"]
 
-from typing import Any, Optional, List
+from importlib.abc import Loader
+from importlib.machinery import ModuleSpec
+from importlib.util import module_from_spec, spec_from_file_location
+from pathlib import Path
+import sys
+from types import ModuleType
+from typing import Any, List, Optional, Union
+
+
+def import_module_from_file(module_file: Union[Path, str]) -> ModuleType:
+    """Import module from file path.
+
+    The name of the imported module is the basename of the specified module
+    file without its extension.
+
+    In addition to being returned by this function, the imported module is
+    loaded into the sys.modules dictionary, allowing for commands such as
+    :code:`from <module> import <class>`.
+
+    Args:
+        module_file: File path of module to import.
+
+    Returns:
+        The imported module.
+    """
+    if not isinstance(module_file, Path):
+        module_file = Path(module_file)
+    module_name = module_file.stem
+
+    module_spec = spec_from_file_location(module_name, module_file)
+
+    if not isinstance(module_spec, ModuleSpec):
+        raise ImportError(f"ModuleSpec not created for module file '{module_file}'")
+    if not isinstance(module_spec.loader, Loader):
+        raise ImportError(f"no loader found for module file '{module_file}'")
+
+    module = module_from_spec(module_spec)
+    sys.modules[module_name] = module
+    module_spec.loader.exec_module(module)
+
+    return module
 
 
 def to_list(x: Optional[Any]) -> List:
