+++ src/test_data/flatfiles/hal_alignment/aln_cache/genome/2bit/genomeA.2bit
