+++ src/test_data/flatfiles/hal_alignment/aln_cache/genome/2bit/genomeB.2bit
