+++ src/test_data/flatfiles/hal_alignment/aln_cache/genome/chrom_sizes/genomeA.chrom.sizes
